--- test/plots/grid-choropleth-dx.ts
@@ -1,46 +1,0 @@
-import * as Plot from "@observablehq/plot";
-import * as d3 from "d3";
-
-export async function gridChoroplethDx() {
-  const [grid, data] = await Promise.all([
-    await d3.csv<any>("data/us-state-grid.csv", d3.autoType).then(gridmap),
-    await d3.csv<any>("data/us-state-population-2010-2019.csv", d3.autoType)
-  ]);
-  const states = data.filter((d) => grid.has(d.State)).map((d) => ({...d, ...grid.get(d.State)}));
-  return Plot.plot({
-    height: 420,
-    x: {
-      axis: null
-    },
-    y: {
-      axis: null
-    },
-    color: {
-      type: "diverging-log",
-      scheme: "piyg"
-    },
-    marks: [
-      Plot.cell(states, {x: "x", y: "y", fill: change, dx: -1.5, dy: -1.5}),
-      Plot.cell(states, {x: "x", y: "y", stroke: "black", dx: 1.5, dy: 1.5}),
-      Plot.text(states, {x: "x", y: "y", text: "key", dy: -6}),
-      Plot.text(states, {
-        x: "x",
-        y: "y",
-        text: (
-          (f) => (d) =>
-            f(change(d) - 1)
-        )(d3.format("+.0%")),
-        dy: 6,
-        fillOpacity: 0.6
-      })
-    ]
-  });
-}
-
-function gridmap(states: {name: string}[]) {
-  return new Map(states.map((state) => [state.name, state]));
-}
-
-function change(d) {
-  return d["2019"] / d["2010"];
-}
